项目改成kotlin, 不过还剩下一部分, 日后再弄

+++ jvm-core/src/main/java/com/gxk/jvm/instruction/constants/BiPushInst.java
@@ -18,7 +18,7 @@ public class BiPushInst implements Instruction {
 
   @Override
   public void execute(Frame frame) {
-    frame.pushInt((int) (this.val));
+    frame.pushInt(this.val);
   }
 
   @Override
